+++ qcal/interface/pygsti/datasets.py
@@ -12,30 +12,25 @@ logger = logging.getLogger(__name__)
 __all__ = 'generate_pygsti_report'
 
 
-def generate_pygsti_dataset(circuits: CircuitSet, fileloc: str):
+def generate_pygsti_dataset(circuits: CircuitSet, save_path: str):
     """Generate a pyGSTi dataset from circuit results.
 
     Args:
         circuits (CircuitSet): measured circuits
-        fileloc (str): location where to save the results.
+        save_path (str): location where to save the results.
     """
-    results_dfs = []
-    for i, circuit in enumerate(circuits):
-        results_dfs.append(
-            pd.DataFrame(
-                [circuit.results.dict], 
-                index=[circuits['pygsti_circuit'][i]]
-            )
-        )
-    results_df = pd.concat(results_dfs)
+    results_df = pd.DataFrame(
+        [circuits['results'].loc[i] for i in range(len(circuits))]
+    )
+    results_df.index = circuits['pygsti_circuit'].values
     results_df = results_df.fillna(0).astype(int).rename(
         columns=lambda col: col + ' count'
     )
 
     logger.info(" Saving the pyGSTi results...")
-    with open(f'{fileloc}dataset.txt', 'w') as f:  # Write the header
+    with open(f'{save_path}dataset.txt', 'w') as f:  # Write the header
         f.write('## Columns = ' + ', '.join(results_df.columns) + "\n")
         f.close()
     results_df.to_csv(
-        f'{fileloc}dataset.txt', sep=' ', mode='a', header=False
+        f'{save_path}dataset.txt', sep=' ', mode='a', header=False
     )